CurrencyDetectorApp/backend/requirements.txt
# Main dependencies for MKD Currency Detector
# Install with: pip install -r requirements.txt

# === Core ML Framework ===
ultralytics>=8.0.134

# === Web Framework ===
fastapi>=0.107.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# === Computer Vision ===
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.26.0
matplotlib>=3.8.0

# === Dataset Management ===
roboflow>=1.0.1

# === Text-to-Speech ===
edge-tts>=6.1.9
playsound3>=1.3.0